% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ACC.R
\name{ACC}
\alias{ACC}
\title{Preparation du fichier ACC}
\usage{
ACC(file, clean_eplet)
}
\arguments{
\item{file}{Chemin vers le fichier .xlsx a importer}

\item{clean_eplet}{Liste issue de la fonction \code{clean_eplet_data()} contenant les tables d'eplets}
}
\value{
Un tableau (data.frame) nettoye et enrichi des eplets
}
\description{
Cette fonction nettoie et prepare le fichier Excel issu du site EpVix (.xls qui doit etre enregistre en .xlsx)
contenant l'analyse des anticorps anti-HLA.
}
\examples{
\dontrun{
case <- ACC("exemple.xlsx", clean_eplet)
}
}
