% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse_ACC_self.R
\name{analyse_ACC_self}
\alias{analyse_ACC_self}
\title{Analyse complete de l'ACC (Allo-anticorps anti-HLA)}
\usage{
analyse_ACC_self(
  acc_file,
  clean_eplet,
  seuil = 1000,
  soi_file = NULL,
  output_excel = TRUE,
  output_name = NULL
)
}
\arguments{
\item{acc_file}{Chemin vers le fichier ACC a analyser.}

\item{clean_eplet}{Liste issue de la fonction \code{clean_eplet_data()} contenant les tables d'eplets}

\item{seuil}{Numeric. Seuil MFI au-dessus duquel un resultat est considere "positif". Default = 1000.}

\item{soi_file}{Chemin vers le fichier contenant les eplets du soi. Default = NULL.}

\item{output_excel}{Logical. Sauvegarder les resultats dans un fichier Excel. Default = TRUE.}

\item{output_name}{Nom du fichier Excel de sortie. Si NULL, un nom par defaut est genere.}
}
\value{
Liste contenant :
\item{case}{Data frame ACC brut}
\item{eplet_soi}{eplets du soi identifies}
\item{seplet_nsallele}{eplets du soi associes a tous les alleles du meme locus}
\item{case_self}{ACC associe aux eplets du soi}
\item{res_total}{Resultats finaux complets}
\item{res_prob}{Resultats finaux les plus probables}
}
\description{
Cette fonction execute l'ensemble du pipeline pour l'analyse des ACC :
lecture des fichiers, identification des eplets du soi, association avec
les alleles, association avec les pairs d'eplets du soi,
calcul des resultats bruts selon un seuil MFI, generation
des resultats finaux et sauvegarde optionnelle en fichier Excel.
}
\examples{
# resultats <- analyse_ACC_self("ACC.xlsx", clean_eplet, seuil = 1000)
}
