% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/associer_acc.R
\name{associer_acc}
\alias{associer_acc}
\title{Associer les eplets du soi avec la recherche d'Ac anti-HLA}
\usage{
associer_acc(case, seplet_nsallele)
}
\arguments{
\item{case}{Data frame contenant les informations du cas (avec colonnes 'allele', 'locus')}

\item{seplet_nsallele}{Data frame contenant les eplets non-soi associes aux alleles (avec colonnes 'allele', 'locus')}
}
\value{
Data frame resultant de la jointure entre le cas et les eplets non-soi
}
\description{
Associer les eplets du soi avec la recherche d'Ac anti-HLA
}
\examples{
# associer_acc(case, seplet_nsallele)
}
