% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculer_resultats.R
\name{calculer_resultats}
\alias{calculer_resultats}
\title{Calculer les resultats bruts des eplets selon un seuil MFI}
\usage{
calculer_resultats(case_f, seuil = 1000)
}
\arguments{
\item{case_f}{Data frame contenant les informations du cas apres association avec les eplets
Doit contenir au minimum les colonnes : 'MFI', 'eplet', 'self', 'exposition', 'status'}

\item{seuil}{Numeric. Seuil MFI au-dessus duquel un resultat est considere "positif". Par defaut 1000.}
}
\value{
Data frame avec les comptes de positifs et negatifs par eplet, auto-filtre sur les positifs uniquement
}
\description{
Calculer les resultats bruts des eplets selon un seuil MFI
}
\examples{
# res <- calculer_resultats(case_f, seuil = 500)
}
