% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_eplet.R
\name{clean_eplet_data}
\alias{clean_eplet_data}
\title{Clean eplet data}
\usage{
clean_eplet_data(raw_data)
}
\arguments{
\item{raw_data}{Un data.frame contenant les donnees brutes issus de la fonction get_epregistry_data.}
}
\value{
Un data.frame nettoye.
}
\description{
Fonction interne utilisee pour nettoyer les donnees des eplets (un eplet et un allele HLA par ligne avec le locus correspondant)
}
