% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/creer_non_self.R
\name{creer_non_self}
\alias{creer_non_self}
\title{Creer la table des eplets du soi sur tous les alleles du meme locus}
\usage{
creer_non_self(eplet_soi, clean_eplet, choix = 1)
}
\arguments{
\item{eplet_soi}{Data frame contenant les eplets du soi avec colonnes 'allele', 'name', 'locus'}

\item{clean_eplet}{Liste issue de la fonction \code{clean_eplet_data()} contenant les tables d'eplets}

\item{choix}{Entier (1, 2 ou 3) pour choisir le sous-ensemble de clean_eplet}
}
\value{
Data frame associant chaque eplet du soi a tous les alleles du meme locus
}
\description{
Creer la table des eplets du soi sur tous les alleles du meme locus
}
\examples{
# creer_non_self(eplet_soi, clean_eplet, choix = 1)
}
