% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/creer_self.R
\name{creer_self}
\alias{creer_self}
\title{Creer la table des eplets du soi}
\usage{
creer_self(soi_file = NULL, case, clean_eplet)
}
\arguments{
\item{soi_file}{Chemin vers le fichier contenant les eplets du soi. Default = NULL.}

\item{case}{Table issue de la fonction \code{ACC()}, contenant les alleles et eplets analyses.}

\item{clean_eplet}{Liste issue de la fonction \code{clean_eplet_data()} contenant les tables d'eplets}
}
\value{
Un tableau des eplets du soi (data.frame)
}
\description{
Cette fonction identifie les eplets du soi a partir :
\itemize{
\item soit du fichier Excel EpVix (feuille 2 si disponible, table des eplets copiee a partir de Epvix),
\item soit automatiquement a partir des donnees issues de la fonction \code{ACC()}.
}
}
\examples{
\dontrun{
eplet_soi <- creer_self(soi_file = NULL, case = case, clean_eplet = clean_eplet)
}
}
