% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generer_resultat_final.R
\name{generer_resultat_final}
\alias{generer_resultat_final}
\title{Generer les resultats finaux a partir des resultats bruts}
\usage{
generer_resultat_final(case_f, res)
}
\arguments{
\item{case_f}{Data frame contenant les informations du cas apres association avec les eplets}

\item{res}{Data frame contenant les resultats bruts generes par \code{calculer_resultats}}
}
\value{
Liste contenant deux data frames :
\item{res_total}{Tous les resultats finaux avec regroupement des alleles et MFI}
\item{res_prob}{Resultats les plus probables (maximum couverture)}
}
\description{
Generer les resultats finaux a partir des resultats bruts
}
\examples{
# resultats <- generer_resultat_final(case_f, res)
}
