% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_eplet.R
\name{get_epregistry_data}
\alias{get_epregistry_data}
\title{Importer et nettoyer les donnees d'eplets HLA}
\usage{
get_epregistry_data(
  local_excel = NULL,
  url = "https://www.epregistry.com.br/databases/ABC"
)
}
\arguments{
\item{local_excel}{Chemin vers un fichier Excel local (.xlsx).}

\item{url}{URL du site EpRegistry (par defaut, https://www.epregistry.com.br/databases/ABC).}
}
\value{
Une liste de tibbles contenant :
\itemize{
\item all_eplets - tous les eplets
\item confirmed - les eplets confirmes
\item exposed - les eplets exposes
}
}
\description{
Ces fonctions permettent d'importer les donnees depuis un fichier Excel local
ou directement depuis le site epregistry.com.br, puis de les nettoyer pour
preparer l'analyse des eplets.
}
\examples{
\dontrun{
clean_eplet <- import_and_clean_eplets()
}
}
